refactor(tests): change how tests are written
